3DInfantPoseEstimation-master/requirements.txt
# 3D Infant Pose Estimation - Requirements
# Python dependencies for the project

# Core ML frameworks
torch>=1.7.0
torchvision>=0.8.0

# Scientific computing
numpy>=1.19.0
scipy>=1.5.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Visualization
matplotlib>=3.3.0

# Utilities
prettytable>=2.0.0

# Note: 
# - Install PyTorch with CUDA support if you have GPU:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# - Or use CPU-only version:
#   pip install torch torchvision

